351 (362) Finished adding payment. moving to rest API (in a new repository)

+++ controllers/admin.js
@@ -1,5 +1,8 @@
 const Product = require("../models/product");
-const {validationResult} = require('express-validator/check')
+const {
+    validationResult
+} = require('express-validator/check')
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     console.log('Add Product page!');
@@ -22,7 +25,7 @@ exports.postAddProduct = (req, res, next) => {
     const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
-    if  (!image) {
+    if (!image) {
         return res.status(422).render("admin/edit-product", {
             pageTitle: "Add Product!",
             path: "/admin/add-product",
@@ -154,6 +157,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl)
                 product.imageUrl = image.path;
             }
             return product.save()
@@ -190,19 +194,27 @@ exports.getProducts = (req, res, next) => {
         });
 }
 
-exports.postDeleteProduct = (req, res, next) => {
-    const prodId = req.body.productId;
-    Product.deleteOne({
-        _id: prodId, 
-        userID: req.user._id
-    })
+exports.deleteProduct = (req, res, next) => {
+    const prodId = req.params.productId;
+    console.log('Post Delete product page!')
+    Product.findById(prodId)
+        .then(product => {
+            console.log(product)
+            if (!product) {
+                return next(new Error('Product not found!'))
+            }
+            fileHelper.deleteFile(product.imageUrl)
+            return Product.deleteOne({ _id: prodId, userId: req.user._id})
+        })
         .then(() => {
             console.log("Destoryed Product!")
-            res.redirect('/admin/products')
+            res.status(200).json({message: 'Success!'});
         })
         .catch(err => {
-            const error = new Error(err)
-            error.httpStatusCode = 500;
-            return next(error);
+            res.status(500).json({message: 'Failed to delete product!'})
+            // const error = new Error(err)
+            // error.httpStatusCode = 500;
+            // return next(error);
         })
+
 }
